backend/pyproject.toml
[project]
name = "leadership-insights-agent-backend"
version = "0.1.0"
description = "Backend for leadership insights agent"
authors = [
    {name = "Dharahas"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "opentelemetry-exporter-otlp-proto-grpc (>=1.40.0,<2.0.0)",
    "unstructured[docx,pdf] (>=0.21.5,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false